% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cbPalette15}
\alias{cbPalette15}
\title{Fifteen-colour palette compatible with colour blindness}
\format{An object of class \code{character} of length 15.}
\source{
\url{http://mkweb.bcgsc.ca/biovis2012/color-blindness-palette.png}
}
\usage{
cbPalette15
}
\description{
A fifteen-colour \href{http://mkweb.bcgsc.ca/brewer/}{Brewer palette} comprehensible
by colour blind viewers.
}
\details{
Note that colour 4 is difficult to distinguish from colour 13 in individuals with tritanopia.
Likewise, colour 7 is difficult to distinguish from colour 3.
You may wish to use \code{cbPalette13 <- cbPalette15[-c(4, 7)]}.
}
\examples{
{
data('cbPalette15')
dev.new(width=8, height=1, units='cm')
par(mar=rep(0, 4))
plot(0, type='n', xlim=c(1, 15), ylim=c(0, 1), axes=FALSE)
points(1:15, rep(0, 15), col=cbPalette15, pch=15)
text(1:15, 0.5, col=cbPalette15)
}

}
\seealso{
\link{cbPalette8}
}
\keyword{datasets}
