% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lexis.R
\name{ColourLexis}
\alias{ColourLexis}
\alias{ColorLexis}
\title{Colour a Lexis plot}
\usage{
ColourLexis(
  values,
  spectrum = hcl.colors(256L, palette = "viridis", alpha = 0.6),
  resolution = attr(values, "effectiveResolution"),
  direction = getOption("ternDirection", 1L),
  legend,
  ...
)

ColorLexis(
  values,
  spectrum = hcl.colors(256L, palette = "viridis", alpha = 0.6),
  resolution = attr(values, "effectiveResolution"),
  direction = getOption("ternDirection", 1L),
  legend,
  ...
)
}
\arguments{
\item{values}{Numeric matrix with four named rows: \code{x}, \code{y} (Cartesian
coordinates of tile centres), \code{z} (value for colour mapping), and \code{down}
(triangle direction: 0 = up, 1 = down).}

\item{spectrum}{Vector of colours to use as a spectrum, or \code{NULL} to use
\code{values["z", ]}.}

\item{resolution}{The number of triangles whose base should lie on the longest
axis of the triangle.  Higher numbers will result in smaller subdivisions and smoother
colour gradients, but at a computational cost.}

\item{direction}{(optional) Integer specifying the direction that the
current ternary plot should point: 1, up; 2, right; 3, down; 4, left.}

\item{legend}{Character vector specifying annotations for colour scale.
If not provided, no colour legend is displayed.
Specify \code{TRUE} to generate automatically, or a single integer to generate
\code{legend} annotations.}

\item{...}{Further arguments to
\code{\link[PlotTools:SpectrumLegend]{SpectrumLegend()}}.}
}
\value{
Called for its side effect (colouring a Lexis plot).  Invisibly
returns \code{NULL}.
}
\description{
Fill the tiles of a Lexis plot with colours determined by a matrix of
values, typically produced by \code{\link[=LexisPointValues]{LexisPointValues()}}.
}
\examples{
LexisPlot(blab = "b", clab = "c")
vals <- LexisPointValues(function(b, c) b - c, resolution = 12)
ColourLexis(vals, legend = TRUE, bty = "n")

}
\seealso{
Other Lexis plotting functions: 
\code{\link{AddToTernary}()},
\code{\link{HoldridgeToXY}()},
\code{\link{LexisContour}()},
\code{\link{LexisPlot}()},
\code{\link{LexisPointValues}()}
}
\author{
\href{https://orcid.org/0000-0001-5660-1727}{Martin R. Smith}
(\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
\concept{Lexis plotting functions}
