% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Holdridge.R, R/TernaryPlot.R
\name{HoldridgeToXY}
\alias{HoldridgeToXY}
\alias{CoordinatesToXY}
\title{Convert user-specified ternary coordinates into X and Y coordinates}
\usage{
HoldridgeToXY(pet, prec)

CoordinatesToXY(coordinates)
}
\arguments{
\item{pet, prec}{Numeric vectors giving \emph{p}otential \emph{e}vapo\emph{t}ranspiration
ratio and annual \emph{prec}ipitation (in mm).}

\item{coordinates}{A list, matrix, data.frame or vector in which each
element (or row) specifies the three coordinates of a point in ternary space.
Each element (or row) will be rescaled such that its entries sum to 100.}
}
\value{
\code{CoordinatesToXY()} returns an array of two rows, corresponding to
the X and Y coordinates of \code{coordinates}.
}
\description{
Accepts various formats of input data; extracts ternary coordinates and
converts to X and Y coordinates.
}
\section{Functions}{
\itemize{
\item \code{HoldridgeToXY()}: Convert from Holdridge coordinates

}}
\author{
\href{https://orcid.org/0000-0001-5660-1727}{Martin R. Smith}
(\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
\keyword{internal}
