% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lexis.R
\name{LexisContour}
\alias{LexisContour}
\title{Add contours to a Lexis plot}
\usage{
LexisContour(
  Func,
  resolution = 96L,
  filled = FALSE,
  legend,
  legend... = list(),
  nlevels = 10,
  levels = pretty(zlim, nlevels),
  zlim,
  color.palette = function(n) {
     hcl.colors(n, palette = "viridis", alpha = 0.6)
 },
  fill.col = color.palette(length(levels) - 1),
  func... = list(),
  ...
)
}
\arguments{
\item{Func}{Function that takes two named arguments corresponding to the
two bounded axes of the active Lexis plot.  For example, if the plot was
created with \code{bRange} and \code{cRange}, \code{Func} should accept arguments
\code{b} and \code{c}.}

\item{resolution}{Integer specifying the number of tile divisions along the
longer axis of the parallelogram.}

\item{filled}{Logical: if \code{TRUE}, draw filled contours.}

\item{legend}{Character or numeric vector of legend labels.  Specify \code{TRUE}
to auto-generate, or a single integer for that many labels.}

\item{legend...}{List of additional arguments to
\code{\link[PlotTools:SpectrumLegend]{SpectrumLegend()}}.}

\item{nlevels}{Integer giving the approximate number of contour levels.}

\item{levels}{Numeric vector of contour levels.}

\item{zlim}{Numeric vector of length 2 giving the z range.}

\item{color.palette}{Function that takes an integer and returns that many
colours.}

\item{fill.col}{Colours for filled contours.}

\item{func...}{List of additional arguments passed to \code{Func()}.}

\item{\dots}{Further arguments passed to \code{\link[graphics:contour]{graphics::contour()}}.}
}
\value{
Invisibly returns a list with elements \code{x}, \code{y}, and \code{z} (the
evaluation grid).
}
\description{
Draw contour lines depicting the value of a function across a Lexis
parallelogram.
}
\examples{
LexisPlot(blab = "b", clab = "c")
LexisContour(function(b, c) b * c, resolution = 36)

}
\seealso{
Other Lexis plotting functions: 
\code{\link{AddToTernary}()},
\code{\link{ColourLexis}()},
\code{\link{HoldridgeToXY}()},
\code{\link{LexisPlot}()},
\code{\link{LexisPointValues}()}
}
\author{
\href{https://orcid.org/0000-0001-5660-1727}{Martin R. Smith}
(\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
\concept{Lexis plotting functions}
