% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lexis.R
\name{LexisPlot}
\alias{LexisPlot}
\title{Create a Lexis diagram}
\usage{
LexisPlot(
  aRange = NULL,
  bRange = NULL,
  cRange = NULL,
  alab = NULL,
  blab = NULL,
  clab = NULL,
  atip = NULL,
  btip = NULL,
  ctip = NULL,
  point = "up",
  clockwise = TRUE,
  xlim = NULL,
  ylim = NULL,
  col = NA,
  padding = 0.08,
  isometric = TRUE,
  grid.lines = 10,
  grid.col = "darkgrey",
  grid.lty = "solid",
  grid.lwd = par("lwd"),
  grid.minor.lines = 4,
  grid.minor.col = "lightgrey",
  grid.minor.lty = "solid",
  grid.minor.lwd = par("lwd"),
  axis.labels = TRUE,
  axis.cex = 0.8,
  axis.col = "black",
  axis.font = par("font"),
  axis.lty = "solid",
  axis.lwd = 1,
  axis.rotate = TRUE,
  axis.tick = TRUE,
  ticks.lwd = axis.lwd,
  ticks.length = 0.025,
  ticks.col = grid.col,
  lab.offset = 0.16,
  lab.cex = 1,
  lab.font = 0,
  lab.col = NULL,
  tip.cex = lab.cex,
  tip.font = 2,
  tip.col = "black",
  free.fun = NULL,
  panel.first = NULL,
  panel.last = NULL,
  ...
)
}
\arguments{
\item{aRange, bRange, cRange}{Numeric vectors of length 2 specifying the
range of each axis, or \code{NULL}.  Exactly two must be non-\code{NULL};
the remaining axis is the "free" (derived) axis.  If all three are
\code{NULL}, the default is \code{bRange = c(0, 1)}, \code{cRange = c(0, 1)}
(i.e., axis \code{a} is free).}

\item{alab, blab, clab}{Character string specifying text with which to label
the corresponding sides of the triangle.
Left or right-pointing arrows are produced by
typing \verb{\\\\U2190} or \verb{\\\\U2192}, or using \code{expression('value' \%->\% '')}.}

\item{atip, btip, ctip}{Character string specifying text to title corners,
proceeding clockwise from the corner specified in \code{point} (default: top).}

\item{point}{Character string specifying the orientation of the ternary plot:
should the triangle point \code{"up"}, \code{"right"}, \code{"down"} or \code{"left"}?
The integers 1 to 4 can be used in place of the character strings.}

\item{clockwise}{Logical specifying the direction of axes.  If \code{TRUE} (the
default), each axis runs from zero to its maximum value in a clockwise
direction around the plot.}

\item{xlim, ylim}{Numeric vectors of length two specifying the minimum and
maximum \emph{x} and \emph{y} limits of the plotted area, to which \code{padding} will be
added.
The default is to display the complete height or width of the plot.
Allows cropping to magnified region of the plot. (See vignette for diagram.)
May be overridden if \code{isometric = TRUE}; see documentation of
\code{isometric} parameter.}

\item{col}{The colour for filling the plot; see
\code{\link[graphics]{polygon}}.}

\item{padding}{Numeric specifying size of internal margin of the plot;
increase if axis labels are being clipped.}

\item{isometric}{Logical specifying whether to enforce an equilateral shape
for the ternary plot.
If only one of \code{xlim} and \code{ylim} is set, the other will be calculated to
maintain an equilateral plot.
If both \code{xlim} and \code{ylim} are set, but have different ranges, then the limit
with the smaller range will be scaled until its range matches that of the
other limit.}

\item{grid.lines}{Integer specifying the number of grid lines to plot.
If \code{axis.labels = TRUE}, this will be used as a hint to \code{pretty()}.}

\item{grid.col, grid.minor.col}{Colours to draw the grid lines. Use a vector
of length three to set different values for each direction.}

\item{grid.lty, grid.minor.lty}{Character or integer vector; line type of
the grid lines. Use a vector of length three to set different values for
each direction.}

\item{grid.lwd, grid.minor.lwd}{Non-negative numeric giving line width of the
grid lines.  Use a vector of length three to set different values for each
direction.}

\item{grid.minor.lines}{Integer specifying the number of minor (unlabelled)
grid lines to plot between each major pair.}

\item{axis.labels}{This can either be a logical value specifying whether
(numerical) annotations are to be made at the tickmarks, or a character or
expression vector of labels to be placed at the tick points,
or a list of length three, with each entry specifying labels to be placed
on each axis in turn.}

\item{axis.cex}{Numeric specifying character expansion (font size)
for axis labels.
Use a vector of length three to set a different value for each direction.}

\item{axis.col, ticks.col, tip.col}{Colours for the axis line,
tick marks and tip labels respectively.
Use a vector of length three to set a different value for each direction.
\code{axis.col = NULL} means to use \code{par('fg')}, possibly specified
inline, and \code{ticks.col = NULL} means to use whatever colour
\code{axis.col} resolved to.}

\item{axis.font}{Font for text. Defaults to \code{par('font')}.}

\item{axis.lty}{Line type for both the axis line and tick marks.
Use a vector of length three to set a different value for each direction.}

\item{axis.lwd, ticks.lwd}{Line width for the axis line and tick marks.
Zero or negative values will suppress the line or ticks.
Use a vector of length three to set different values for each axis.}

\item{axis.rotate}{Logical specifying whether to rotate axis labels
to parallel grid lines, or numeric specifying custom rotation for each axis,
to be passed as \code{srt} parameter to \code{text()}.  Expand margins or set
\code{par(xpd = NA)} if labels are clipped.}

\item{axis.tick}{Logical specifying whether to mark the axes with tick marks.}

\item{ticks.length}{Numeric specifying distance that ticks should extend
beyond the plot margin.  Also affects position of axis labels, which are
plotted at the end of each tick.
Use a vector of length three to set a different length for each direction.}

\item{lab.offset}{Numeric specifying distance between midpoint of axis label
and the axis. The default value is given in the 'Usage' section; a value
of \code{0} will position the axis label directly on the axis.
Increase \code{padding} if labels are being clipped.
Use a vector of length three to specify a different offset for each label.}

\item{lab.cex, tip.cex}{Numeric specifying character expansion (font size)
for axis labels.
Use a vector of length three to specify a different value for each direction.}

\item{lab.font, tip.font}{Numeric specifying font style (Roman, bold, italic,
bold-italic) for axis titles.
Use a vector of length three to set a different font for each direction.}

\item{lab.col}{Character vector specifying colours for axis labels. Use a
vector of length three to specify a different colour for each label.}

\item{free.fun}{Optional function used to compute labels for the free
(derived) axis.
Called with named arguments matching the two bounded axes.
For example, if \code{aRange} and \code{bRange} are specified (so \code{c} is free),
pass \code{free.fun = function(a, b) b - a} to label the free axis with
\eqn{b - a}.  If \code{NULL} (the default), labels show the ternary
complement \eqn{k - a' - b'} where primes denote shifted coordinates
and \eqn{k = \Delta a + \Delta b}.}

\item{panel.first}{An expression to be evaluated after the plot axes are
set up but before any plotting takes place.
This can be useful for drawing backgrounds, e.g. with \code{\link[=ColourTernary]{ColourTernary()}}
or \code{\link[=HorizontalGrid]{HorizontalGrid()}}.
Note that this works by lazy evaluation: passing this argument from other
plot methods may well not work since it may be evaluated too early.}

\item{panel.last}{An expression to be evaluated after plotting has taken
place but before the axes and box are added.  See the comments about
\code{panel.first}.}

\item{...}{Additional parameters to \code{\link[graphics]{plot}}.}
}
\value{
A list of class \code{"LexisPlot"} containing plot parameters, returned
invisibly.
}
\description{
\code{LexisPlot()} creates a blank Lexis diagram (parallelogram-shaped ternary
plot) onto which data can be plotted using the \code{\link[=AddToLexis]{AddToLexis()}} family of
functions.
}
\details{
A Lexis diagram has three axes subject to the constraint
\eqn{a + b + c = \textrm{constant}}{a + b + c = constant}.  The user
specifies ranges for two of the three axes; the third is derived.
}
\examples{
LexisPlot()
LexisPoints(b = c(0.2, 0.4), c = c(0.3, 0.6))

# Age-Period-Cohort example (c free, with Cohort labels)
LexisPlot(aRange = c(0, 100), bRange = c(1900, 2000),
          alab = "Age", blab = "Period", clab = "Cohort",
          free.fun = function(a, b) b - a)
LexisPoints(a = c(20, 50), b = c(1950, 1980))

}
\seealso{
Other Lexis plotting functions: 
\code{\link{AddToTernary}()},
\code{\link{ColourLexis}()},
\code{\link{HoldridgeToXY}()},
\code{\link{LexisContour}()},
\code{\link{LexisPointValues}()}
}
\author{
\href{https://orcid.org/0000-0001-5660-1727}{Martin R. Smith}
(\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
\concept{Lexis plotting functions}
