% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Polygons.R
\name{Polygon-Geometry}
\alias{Polygon-Geometry}
\alias{PolygonArea}
\title{Polygon geometry
These deprecated functions have been moved to "PlotTools".}
\usage{
PolygonArea(...)
}
\arguments{
\item{\dots}{Parameters to PlotTools function}
}
\description{
Polygon geometry
These deprecated functions have been moved to "PlotTools".
}
