% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coordinates.R
\name{TernaryXRange}
\alias{TernaryXRange}
\alias{TernaryYRange}
\title{X and Y coordinates of ternary plotting area}
\usage{
TernaryXRange(direction = getOption("ternDirection", 1L))

TernaryYRange(direction = getOption("ternDirection", 1L))
}
\arguments{
\item{direction}{(optional) Integer specifying the direction that the
current ternary plot should point: 1, up; 2, right; 3, down; 4, left.}
}
\value{
\code{TernaryXRange()} and \code{TernaryYRange()} return the minimum and
maximum X or Y coordinate of the area in which a ternary plot is drawn,
oriented in the specified direction.
Because the plotting area is a square, the triangle of the ternary plot
will not occupy the full range in one direction.
Assumes that the defaults have not been overwritten by specifying \code{xlim} or
\code{ylim}.
}
\description{
X and Y coordinates of ternary plotting area
}
\section{Functions}{
\itemize{
\item \code{TernaryYRange()}: Returns the minimum and maximum Y coordinate for a
ternary plot in the specified direction.

}}
\seealso{
Other plot limits:
\code{\link[=OutsidePlot]{OutsidePlot()}}
}
\author{
\href{https://orcid.org/0000-0001-5660-1727}{Martin R. Smith}
(\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
\concept{plot limits}
