% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annotate.R
\name{Annotate}
\alias{Annotate}
\title{Annotate individual points on a ternary diagram}
\usage{
Annotate(
  coordinates,
  labels,
  side,
  line.col = col,
  lty = par("lty"),
  lwd = par("lwd"),
  col = par("col"),
  font = par("font"),
  ...
)
}
\arguments{
\item{coordinates}{A list, matrix, data.frame or vector in which each
element (or row) specifies
the three coordinates of a point in ternary space.}

\item{labels}{Character vector of same length as \code{coordinates}
specifying text with which to annotate each point.}

\item{side}{Optional integer vector specifying which side of the ternary
plot each point should be labelled on. Points labelled 0 will not be
annotated (but still require an entry in \code{annotation}).}

\item{line.col, lty, lwd}{parameters to \code{\link[=segments]{segments()}}.}

\item{col, font}{parameters to \code{\link[=text]{text()}}.}

\item{\dots}{Further parameters to \code{\link[=text]{text()}} and \code{\link[=segments]{segments()}}.}
}
\description{
Annotate individual points on a ternary diagram
}
\examples{
data("Seatbelts")
seats <- c("drivers", "front", "rear")
seat <- Seatbelts[month.abb \%in\% "Oct", seats]
TernaryPlot(alab = seats[1], blab = seats[2], clab = seats[3])
TernaryPoints(seat, pch = 16, cex = 0.5)
Annotate(seat[1, ], side = 1)
Annotate(seat, labels = 1969:1984)

 
}
\author{
\href{https://orcid.org/0000-0001-5660-1727}{Martin R. Smith}
(\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
