% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Contours.R
\name{ColourTernary}
\alias{ColourTernary}
\title{Colour a ternary plot according to the output of a function}
\usage{
ColourTernary(values, spectrum = viridisLite::viridis(256L, alpha = 0.6),
  resolution = sqrt(ncol(values)),
  direction = getOption("ternDirection"))
}
\arguments{
\item{values}{Numeric vector specifying the values associated with each point,
generated using \code{[TernaryPointValues]}.}

\item{spectrum}{Vector of colours to use as a spectrum.}

\item{resolution}{Number of rows into which the larger triangle should be subdivided.}

\item{direction}{(optional) Integer specifying the direction that the current ternary plot
should point: 1, up; 2, right; 3, down; 4, left.}
}
\description{
Colour a ternary plot according to the output of a function
}
\author{
Martin R. Smith
}
