% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TernaryPlot.R
\name{CoordinatesToXY}
\alias{CoordinatesToXY}
\title{Convert user-specified ternary coordinates into X and Y coordinates}
\usage{
CoordinatesToXY(coordinates)
}
\arguments{
\item{coordinates}{A list, matrix, data.frame or vector in which each
element (or row) specifies
the three coordinates of a point in ternary space.}
}
\value{
\code{CoordinatesToXY()} returns an array of two rows, corresponding to
the X and Y coordinates of \code{coordinates}.
}
\description{
Accepts various formats of input data; extracts ternary coordinates and
converts to X and Y coordinates.
}
\author{
\href{https://orcid.org/0000-0001-5660-1727}{Martin R. Smith}
(\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
\keyword{internal}
