% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetRegion.R
\name{.SetRegion}
\alias{.SetRegion}
\title{Set plotting region}
\usage{
.SetRegion(region, prettify = NA_integer_)
}
\arguments{
\item{region}{List of length three specifying the sub-region of each ternary
axis to be drawn; or a set of coordinates in a format accepted by
\code{\link[=TernaryPoints]{TernaryPoints()}}.
The plotted region will correspond to the smallest equilateral triangle
that encompasses the specified ranges or coordinates.}

\item{prettify}{If numeric, the plotting region will be expanded to allow
grid lines to be produced with \code{pretty(n = prettify)}. If \code{NA}, the
smallest region encompassing \code{region} will be used.}
}
\value{
\code{SetRegion()} returns the value of \code{options(ternRegion = region)}.
}
\description{
Sets the region of the ternary plot being drawn.
Usually called from within \code{TernaryPlot()}; everyday users are unlikely to
need to call this function directly.
}
\examples{
# XY Coordinates under original plotting region
TernaryToXY(c(1, 2, 3))
previous <- SetRegion(ternRegion20)

# New region options set
getOption("ternRegion")

# Coordinates under new plotting region
TernaryToXY(c(1, 2, 3))

# Restore previous setting
options(previous)
getOption("ternRegion")
}
\author{
\href{https://orcid.org/0000-0001-5660-1727}{Martin R. Smith}
(\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
\keyword{internal}
