% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{holdridge}
\alias{holdridge}
\title{Random sample of points for Holdridge plotting}
\format{
An object of class \code{data.frame} with 39 rows and 4 columns.
}
\usage{
holdridge
}
\description{
A stratified random sampling (average of 100 points) using
a global mapping of Holdridge’s scheme.
}
\examples{
data('holdridge', package = 'Ternary')
head(holdridge)
}
\author{
James Lee Tsakalos
}
\keyword{datasets}
