% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Holdridge.R
\name{HoldridgeHypsometricCol}
\alias{HoldridgeHypsometricCol}
\title{Convert a point in evapotranspiration-precipitation space to an appropriate
cross-blended hypsometric colour}
\usage{
HoldridgeHypsometricCol(pet, prec, opacity = NA)
}
\arguments{
\item{pet}{Numeric vectors giving \emph{p}otential \emph{e}vapo\emph{t}ranspiration
ratio and annual \emph{prec}ipitation (in mm).}

\item{prec}{Numeric vectors giving \emph{p}otential \emph{e}vapo\emph{t}ranspiration
ratio and annual \emph{prec}ipitation (in mm).}
}
\value{
Character vector listing RGBA values corresponding to each pet-prec
value.
}
\description{
Palette basis: https://www.shadedrelief.com/hypso/hypso.html
}
\author{
\href{https://orcid.org/0000-0001-5660-1727}{Martin R. Smith}
(\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
