% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coordinates.R
\name{OutsidePlot}
\alias{OutsidePlot}
\title{Is a point in the plotting area?}
\usage{
OutsidePlot(x, y)
}
\arguments{
\item{x, y}{Vectors of \emph{x} and \emph{y} coordinates of points}
}
\value{
Logical vector specifying whether each pair of \emph{x} and \emph{y} coordinates
corresponds to a point outside the plotted ternary diagram.
}
\description{
Is a point in the plotting area?
}
\author{
Martin R. Smith
}
