% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Polygons.R
\name{Polygon-Geometry}
\alias{Polygon-Geometry}
\alias{PolygonArea}
\alias{PolygonCenter}
\alias{PolygonCentre}
\alias{GrowPolygon}
\title{Polygon geometry}
\usage{
PolygonArea(...)

PolygonCenter(...)

PolygonCentre(...)

GrowPolygon(...)
}
\arguments{
\item{\dots}{Parameters to PlotTools function}
}
\description{
These functions have moved to "\href{https://ms609.github.io/PlotTools}{PlotTools}" and are deprecated here.
}
\keyword{internal}
