% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coordinates.R
\name{ReflectedEquivalents}
\alias{ReflectedEquivalents}
\title{Reflected equivalents of points outside the ternary plot}
\usage{
ReflectedEquivalents(x, y, direction = getOption("ternDirection"))
}
\arguments{
\item{x, y}{Vectors of \emph{x} and \emph{y} coordinates of points.}

\item{direction}{(optional) Integer specifying the direction that the current ternary plot
should point: 1, up; 2, right; 3, down; 4, left.}
}
\value{
A list of the \emph{x}, \emph{y} coordinates of the points produced if
the given point is reflected across each of the edges or corners.
}
\description{
To avoid edge effects, it may be desirable to add the value of a point
within a ternary plot with the value of its 'reflection' across the nearest
axis or corner.
}
