% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ternary-package.R
\docType{package}
\name{Ternary-package}
\alias{Ternary}
\alias{Ternary-package}
\title{Ternary: Create Ternary Plots}
\description{
Plots ternary diagrams (simplex plots / Gibbs triangles) using the
  standard graphics functions.
  An alternative to 'ggtern', which uses the 'ggplot2' family of plotting 
  functions.
}
\details{
For usage instructions, please consult the guidance article
\href{https://ms609.github.io/Ternary/articles/Ternary.html}{online} or by
typing \code{vignette('Ternary', 'Ternary')} at the R console.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ms609.github.io/Ternary/}
  \item \url{https://github.com/ms609/Ternary/}
  \item Report bugs at \url{https://github.com/ms609/Ternary/issues/}
}

}
\author{
\strong{Maintainer}: Martin R. Smith \email{martin.smith@durham.ac.uk} (\href{https://orcid.org/0000-0001-5660-1727}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Lilian Sanselme [contributor]
}

}
\keyword{internal}
