% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Contours.R
\name{TernaryContour}
\alias{TernaryContour}
\title{Add contours to a ternary plot}
\usage{
TernaryContour(Func, resolution = 96L,
  direction = getOption("ternDirection"), ...)
}
\arguments{
\item{Func}{Function taking the parameters \code{a}, \code{b} and \code{c}, which evaluates to a numeric
whose value should be depicted.}

\item{resolution}{The number of triangles whose base should lie on the longest
axis of the triangle.  Higher numbers will result in smaller subdivisions and smoother
colour gradients, but at a computational cost.}

\item{direction}{(optional) Integer specifying the direction that the current ternary plot
should point: 1, up; 2, right; 3, down; 4, left.}

\item{\dots}{Further parameters to pass to `\link[graphics]{contour}.}
}
\description{
Draws contour lines to depict the value of a function in ternary space.
}
\seealso{
Other contour plotting functions: \code{\link{ColourTernary}},
  \code{\link{TernaryDensityContour}},
  \code{\link{TernaryPointValues}}
}
\author{
Martin R. Smith (\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
\concept{contour plotting functions}
