% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Contours.R
\name{TernaryContour}
\alias{TernaryContour}
\title{Add contours to a ternary plot}
\usage{
TernaryContour(
  Func,
  resolution = 96L,
  direction = getOption("ternDirection", 1L),
  within = NULL,
  filled = FALSE,
  legend,
  legend... = list(),
  nlevels = 10,
  levels = pretty(zlim, nlevels),
  zlim,
  color.palette = function(n) viridisLite::viridis(n, alpha = 0.6),
  fill.col = color.palette(length(levels) - 1),
  ...
)
}
\arguments{
\item{Func}{Function taking vectors of coordinates \code{a}, \code{b} and \code{c}, which
returns a numeric vector whose value at each coordinate will be depicted.}

\item{resolution}{The number of triangles whose base should lie on the longest
axis of the triangle.  Higher numbers will result in smaller subdivisions and smoother
colour gradients, but at a computational cost.}

\item{direction}{(optional) Integer specifying the direction that the current ternary plot
should point: 1, up; 2, right; 3, down; 4, left.}

\item{within}{List or matrix of \emph{x, y} coordinates within which contours
should be evaluated, in any format supported by
\code{\link[grDevices:xy.coords]{xy.coords(x = within)}}.
If \code{NULL}, defaults to a region slightly smaller than the ternary plot.
The \verb{$hull} entry generated by \code{TriangleInHull()} may also be used.}

\item{filled}{Logical; if \code{TRUE}, contours will be filled
(using \code{\link[graphics]{.filled.contour}()}.).}

\item{legend}{Character vector specifying annotations for colour scale.
If not provided, no colour legend is displayed.
Specify \code{TRUE} to generate automatically, or a single integer to generate
\code{legend} annotations.}

\item{legend...}{List of additional parameters to send to
\code{\link[=SpectrumLegend]{SpectrumLegend()}}.}

\item{nlevels, levels, zlim, \dots}{parameters to pass to
\code{\link[graphics]{contour}()}.}

\item{color.palette}{parameters to pass to
\code{\link[graphics]{.filled.contour}()}.}

\item{fill.col}{Sent as \code{col} parameter to
\code{\link[graphics]{.filled.contour}()}.
Computed from \code{color.palette} if not specified.}
}
\value{
\code{TernaryContour()}  invisibly returns a list containing:
\itemize{
\item \code{x},\code{y}: the Cartesian coordinates of each evaluated point;
\item \code{z}: The value of \code{Func()} at each coordinate.
}
}
\description{
Draws contour lines to depict the value of a function in ternary space.
}
\examples{
FunctionToContour <- function (a, b, c) {
  a - c + (4 * a * b) + (27 * a * b * c)
}

# Set up plot
originalPar <- par(mar = rep(0, 4))
TernaryPlot(alab = "a", blab = "b", clab = "c")
values <- TernaryPointValues(FunctionToContour, resolution = 24L)
ColourTernary(values,
  legend = signif(seq(max(values), min(values), length.out = 4), 2),
   bty = "n")
TernaryContour(FunctionToContour, resolution = 36L)

# Note that FunctionToContour is sent a vector.
# Instead of
BadMax <- function (a, b, c) {
  max(a, b, c)
}

# Use
GoodMax <- function (a, b, c) {
  pmax(a, b, c)
}
TernaryPlot(alab = "a", blab = "b", clab = "c")
ColourTernary(TernaryPointValues(GoodMax))
TernaryContour(GoodMax)

# Or, for a generalizable example,
GeneralMax <- function (a, b, c) {
  apply(rbind(a, b, c), 2, max)
}
TernaryPlot(alab = "a", blab = "b", clab = "c")
# Fill the contour areas, rather than using tiles
TernaryContour(GeneralMax, filled = TRUE,
               legend = c("Max", "Min"), bty = "n",
               fill.col = viridisLite::viridis(14, alpha = 0.6))
# Re-draw edges of plot triangle over fill
TernaryPolygon(diag(3))

# Restore plotting parameters
par(originalPar)
}
\seealso{
Other contour plotting functions: 
\code{\link{ColourTernary}()},
\code{\link{TernaryDensityContour}()},
\code{\link{TernaryPointValues}()}
}
\author{
\href{https://orcid.org/0000-0001-5660-1727}{Martin R. Smith}
(\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
\concept{contour plotting functions}
