% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TernaryPlot.R
\name{TernaryCoords}
\alias{TernaryCoords}
\title{XY to Ternary}
\usage{
TernaryCoords(abc, b_coord = NULL, c_coord = NULL)
}
\arguments{
\item{abc}{A vector of length three giving the position on a ternary plot.
\code{c(100, 0, 0)} will plot in the top corner; \code{c(0, 100, 0)}
will plot in the bottom corner; \code{c(0, 0, 1)} will plot in
the right-hand corner.
Alternatively, the a coordinate can be specified as the first parameter,
in which case the b and c coordinates must be specified via \code{b_coord}
and \code{c_coord}}

\item{b_coord}{The b coordinate, if \code{abc} is a single number.}

\item{c_coord}{The c coordinate, if \code{abc} is a single number.}
}
\value{
A vector of length two that converts the coordinates given in \code{abc}
into cartesian (x, y) coordinates, to be overlaid on a ternary plot
with corners at (0, -1/2), (0, 1/2), (sin(pi/3), 0)
}
\description{
XY to Ternary
}
\author{
Martin R. Smith
}
