% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Contours.R
\name{TernaryPointValues}
\alias{TernaryPointValues}
\title{Value of a function at regularly spaced points}
\usage{
TernaryPointValues(Func, resolution = 96L,
  direction = getOption("ternDirection"))
}
\arguments{
\item{direction}{(optional) Integer specifying the direction that the current ternary plot
should point: 1, up; 2, right; 3, down; 4, left.}
}
\value{
A matrix whose rows correspond to:

\strong{x}, \strong{y}: co-ordinates of the centres of smaller triangles

\strong{z}: The value of \code{Func(x, y)}

\strong{down}: \code{0} if the triangle concerned points upwards (or right), \code{1} otherwise
}
\description{
Value of a function at regularly spaced points
}
\author{
Martin R. Smith
}
