% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TernaryPlot.R
\name{TernaryXRange}
\alias{TernaryXRange}
\alias{TernaryYRange}
\title{X and Y coordinates of ternary plotting area}
\usage{
TernaryXRange(direction = getOption("ternDirection"))

TernaryYRange(direction = getOption("ternDirection"))
}
\arguments{
\item{direction}{Integer specifying the direction that the current ternary plot 
should point: 1, up; 2, right; 3, down; 4, left.}
}
\value{
Returns the minimum and maximum X coordinate for a ternary plot,
oriented in the specified direction.
}
\description{
X and Y coordinates of ternary plotting area
}
\section{Functions}{
\itemize{
\item \code{TernaryYRange}: Returns the minimum and maximum Y coordinate for a ternary plot in the specified direction.
}}

\author{
Martin R. Smith
}
