% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Contours.R
\name{TriangleCentres}
\alias{TriangleCentres}
\title{Coordinates of triangle mid-points}
\usage{
TriangleCentres(resolution = 48L,
  direction = getOption("ternDirection"))
}
\arguments{
\item{resolution}{The number of triangles whose base should lie on the longest
axis of the triangle.  Higher numbers will result in smaller subdivisions and smoother
colour gradients, but at a computational cost.}

\item{direction}{(optional) Integer specifying the direction that the current ternary plot
should point: 1, up; 2, right; 3, down; 4, left.}
}
\description{
Coordinates of triangle mid-points
}
